Get perceptual-differences within palette

Get perceptual-differences within palette

pev_data_separation(pal, method = "cie2000", include_cvd = TRUE)

Arguments

pal
character (coerceable to vector of hex-codes), describes a discrete color palette. Interpolation within a palette is not required to be a well-formed idea.

method
character method to use for comparison, passed to farver::compare_color(). One of: "euclidean", "cie1976", "cie94", "cie2000", or "cmc".

include_cvd
logical, indicates to include differences for color-vision deficiency

Value

tbl_df with variables cvd, color_a, color_b, difference

See also

Examples

  library("colorspace")
  pal <- qualitative_hcl(n = 5)
  pev_data_separation(pal)

#> # A tibble: 100 x 4
#>    cvd   color_a color_b difference
#>    <chr> <fct>   <chr>        <dbl>
#>  1 none  #E16A86 #E16A86        0  
#>  2 none  #E16A86 #AA9000       48.6
#>  3 none  #E16A86 #00AA5A       71.4
#>  4 none  #E16A86 #00A6CA       60.3
#>  5 none  #E16A86 #B675E0       23.0
#>  6 none  #AA9000 #E16A86       48.6
#>  7 none  #AA9000 #AA9000        0  
#>  8 none  #AA9000 #00AA5A       30.4
#>  9 none  #AA9000 #00A6CA       47.7
#> 10 none  #AA9000 #B675E0       63.9
#> # … with 90 more rows