Create HCL parameter-set

This follows the colorspace scheme.

pev_hcl(type = c("qualitative", "sequential", "diverging"), h1,
  h2 = NULL, c1, c2 = NULL, l1, l2 = NULL, p1 = NULL, p2 = NULL,
  cmax = NULL, fixup = TRUE)

Arguments

type
character indicating type of HCL palette.

h1
hue value in the HCL color description, has to be in [0, 360].

h2
hue value in the HCL color description, has to be in [0, 360].

c1
chroma value in the HCL color description.

c2
chroma value in the HCL color description.

l1
luminance value in the HCL color description.

l2
luminance value in the HCL color description.

p1
control parameter determining how chroma and luminance should be increased (1 = linear, 2 = quadratic, etc.).

p2
control parameter determining how chroma and luminance should be increased (1 = linear, 2 = quadratic, etc.).

cmax
Maximum chroma value in the HCL color description.

fixup
logical. Should the color be corrected to a valid RGB value?

Value

Object with S3 class pev_hcl

Examples

 pev_hcl(h1 = 0, h2 = 360, c1 = 60, l1 = 60)