Get perceptual-separation within discrete-palette

Get perceptual-separation within discrete-palette

pev_data_separation(.fdisc, n = NULL, method = "cie2000",
  include_cvd = TRUE)

Arguments

.fdisc
object that can be coerced to pev_fdisc, when called with an integer, returns the corresponding (hex-code) values.

n
integer number of colors to use for unbounded palette-functions

method
character, describes how the domain of the continuous palette is to be discretized, can be "post" or "panel" (using a fencing analogy).

include_cvd
logical or character, indicates to include data for for color-vision deficiency. Possible character values: "none", "protan", "deutan", "tritan".

Value

tbl_df with variables cvd, hex_a, hex_b, distance

See also

Examples

  pev_fcont("Dynamic") %>%
    pev_fdisc(n = 7, method = "panel") %>%
    pev_data_separation()

#> # A tibble: 484 x 6
#>    cvd       i i_ref hex     hex_ref distance
#>    <chr> <int> <int> <chr>   <chr>      <dbl>
#>  1 none      1     1 #E19896 #E19896      0  
#>  2 none      1     2 #E19896 #E393B8     12.5
#>  3 none      1     3 #E19896 #D497D3     19.4
#>  4 none      1     4 #E19896 #B3A2E1     24.8
#>  5 none      1     5 #E19896 #82AFDE     34.7
#>  6 none      1     6 #E19896 #4CB9CC     52.4
#>  7 none      1     7 #E19896 #3BBEAE     48.8
#>  8 none      1     8 #E19896 #64BC8C     49.0
#>  9 none      1     9 #E19896 #91B66F     42.9
#> 10 none      1    10 #E19896 #B6AD65     32.7
#> # … with 474 more rows