Get separation within discrete-palette

Get separation within discrete-palette

pev_data_separation(.fdisc, method = "cie2000", include_cvd = TRUE)

Arguments

.fdisc
object that can be coerced to pev_fdisc, when called with an integer, returns the corresponding (hex-code) values.

method
character method to use for comparison, passed to farver::compare_color(). One of: "euclidean", "cie1976", "cie94", "cie2000", or "cmc".

include_cvd
logical, indicates to include data for for color-vision deficiency

Value

tbl_df with variables cvd, hex_a, hex_b, distance

See also

Examples

  fcont <- pev_fcont("Dynamic")
  fdisc <- pev_fdisc(fcont, n = 7, method = "panel")
  pev_data_separation(fdisc)

#> # A tibble: 196 x 5
#>    cvd   index_a hex_a   hex_b   distance
#>    <chr>   <int> <chr>   <chr>      <dbl>
#>  1 none        1 #E396A0 #E396A0      0  
#>  2 none        1 #E396A0 #D796D0     14.9
#>  3 none        1 #E396A0 #9FA8E2     26.4
#>  4 none        1 #E396A0 #4CB9CC     53.6
#>  5 none        1 #E396A0 #50BE9B     53.0
#>  6 none        1 #E396A0 #97B56C     45.3
#>  7 none        1 #E396A0 #CBA56E     27.1
#>  8 none        2 #D796D0 #E396A0     14.9
#>  9 none        2 #D796D0 #D796D0      0  
#> 10 none        2 #D796D0 #9FA8E2     18.3
#> # … with 186 more rows